Usie ImmerReducer to make redux easier and typesafe
Redux is great, but it requires a lot of boilerplate code which slows down development, and is not type safe by default.
By using ImmerReducer, we are able to retain all the benefits of redux, with a more familiar approach to doing state mutations.
In addition, there is no longer the need to write actions by hand, as they are generated by the ImmerReducer library from methods of reducer classes.